Restructure test harness.
The test harness for a particular test subdirectory will consist of a
wrapper bash script in addition to the core test program (e.g.
sysaclls.c).  

The test script will be responsible for:
- system config file backup, modification and restore (e.g. auditd
  configuration and audit rules)
- reading a test data file and iterating through individual testcases
- tallying tests passed, failed and skipped
- handling test output levels, which are as follows:
    - normal: pass/fail
    - verbose: pass/fail, debug output for failed tests
    - debug: pass/fall, debug output for all tests

The core program will:
- lookup the testcase routine corresponding to the provided test name
- execute the specified testcase according to the "success" flag
- print all debug info to stdout
- return one of the following values:
    - 0  test completed as expected
    - 1  test completed with unexpected result
    - 2  test did not complete due to error

Test users will be created via a setup script, which will be run on
installation of the test suite.

I've made the majority of the necessary changes in the utils and
syscalls subdirectories.  Will need to modify the core programs in the
other subdirectories accordingly.



git-svn-id: http://audit-test.svn.sourceforge.net/svnroot/audit-test/trunk/tests@487 df539fd5-5ffc-0310-83f5-f045c3a6a7d5