Add a default PATH to run.bash so it doesn't depend on the tester's environment
For directories using the run.bash test harness, move the "run" target to
rules.mk to simplify Makefiles.

Change the subdirectory traversal mechanism.  Now it's possible to simply
specify, for example, $(TOPDIR)/utils/testutils.a in a dependency list, and it
will be built correctly.  There is no need to additionally list $(TOPDIR)/utils
in SUB_DIRS, in fact that is a (relatively harmless) mistake.

Update the dependency checking in rules.mk so that it's possible to run "make
verify" at any level to make sure the system is prepared to run the tests.

 Makefile           |   20 ++---
 fail-safe/Makefile |    9 --
 rules.mk           |  190 ++++++++++++++++++++++++++++-------------------------
 run.bash           |    8 +-
 syscalls/Makefile  |   11 ---
 5 files changed, 121 insertions(+), 117 deletions(-)



git-svn-id: http://audit-test.svn.sourceforge.net/svnroot/audit-test/trunk/tests@580 df539fd5-5ffc-0310-83f5-f045c3a6a7d5