Completely rewrite the log parser to correctly handle quoted values and keys with spaces.
Use merging for parsing as well as combining records.  To our disgruntlement,
this is necessary because some audit records actually have different fields with
the same key on the same line!

Add special handling for msg_time and msg_seq.  These are models for any further
special handling that might be added in the future.

Add section to man-page describing expression syntax.



git-svn-id: http://audit-test.svn.sourceforge.net/svnroot/audit-test/trunk/tests@705 df539fd5-5ffc-0310-83f5-f045c3a6a7d5