Instead of handling keys-with-spaces generically, handle the known abberations specifically.  This allows us to more easily and correctly handle extra text in the audit record, storing it in a special field creatively named "extra_text"
In addition to handling keys-with-spaces specifically, replace the
whitespace in them with underscores to make the command-line and
output sane.

Allow = and ~ to be used on the cmdline as synonym operators for ==
and =~ respectively.  In particular, using = makes cutting and pasting
easier.

Here is an example demonstrating all the changes in this commit.
First the original raw record, then the parsed and formatted record:

$ augrep -m1 type=DAEMON_START --raw
type=DAEMON_START msg=audit(1133214700.153:792)
    auditd start, ver=1.0.3, format=raw, auid=4294967295, 
    auditd pid=1698

$ augrep -m1 type=DAEMON_START
type=DAEMON_START msg=audit(1133214700.153:792) 
    extra_text="auditd start," ver=1.0.3, format=raw, auid=4294967295,
    auditd_pid=1698



git-svn-id: http://audit-test.svn.sourceforge.net/svnroot/audit-test/trunk/tests@743 df539fd5-5ffc-0310-83f5-f045c3a6a7d5