fail-safe/max_log_file rotate was failing.  It appears to be that auditd was not accepting records when write_records was called, so not all the records were logged and the max size wasn't triggered.
One problem was that start_auditd was timing out after 2 seconds when
it was supposed to be waiting for 20.  Fix the magnitude issue and
split the difference to 10 seconds.

Otherwise, harden the tests in general and improve their debugging
output so failures can be more easily pinpointed in the future.



git-svn-id: http://audit-test.svn.sourceforge.net/svnroot/audit-test/trunk/tests@927 df539fd5-5ffc-0310-83f5-f045c3a6a7d5