v2 Fix ipv6 address manipulation
Message-Id: <7a69edf81c88e9e078a632c4aab33b316164c213.1300049990.git.linda.knippers@hp.com>
In-Reply-To: <cover.1300049990.git.linda.knippers@hp.com>
References: <cover.1300049990.git.linda.knippers@hp.com>
From: Linda Knippers <linda.knippers@hp.com>
Date: Sun, 13 Mar 2011 16:56:57 -0400
Subject: [PATCH 3/3] v2 Fix ipv6 address manipulation

v2 addresses Paul's comments and includes his rework of the
prefix code.

The code that extracts ipv6 interfaces and addresses was trying to
only look for interfaces with the appropriate address prefix but
was doing it wrong and not getting the interfaces or the addresses
right.  I've only tested this on a system with one active ethernet
interface but it works there and seems to be the right fix for
systems with multiple interfaces.

Signed-off-by: Linda Knippers <linda.knippers@hp.com>