Update the rpm build for git and RHEL6
Message-Id: <07e0e383bd03db863516a34964beac5317b5068a.1300329492.git.linda.knippers@hp.com>
In-Reply-To: <cover.1300329492.git.linda.knippers@hp.com>
References: <cover.1300329492.git.linda.knippers@hp.com>
From: Linda Knippers <linda.knippers@hp.com>
Date: Wed, 16 Mar 2011 22:12:50 -0400
Subject: [PATCH 1/4] Update the rpm build for git and RHEL6

The 'dist' and 'rpm' targets of the Makefile assumed svn and
and older style of rpmbuild.  Updated for git, using the
commit string for the version, and the current style of rpmbuild,
which puts things in a different location.

The dependency building part (which is really in rules.mk) still
needs work so that part is hacked right now.  See audit-test
project bug #3218177.

Signed-off-by: Linda Knippers <linda.knippers@hp.com>