Update for current netlabelctl behavior
The RHEL6 netlabelctl command displays different output for
the unlbl list command so we need to account for that.

Also, sometimes the test gets an EBUSY when it tries to remove
something from the netlabel configuration but retrying seems to
work around it.  I'm not sure why we're getting EBUSY but will live
with it for now.  With these changes this test now passes.
If we're going to live with the retries indefinitely, we should
probably create a function that does all the work.