v2: Basic network test support
v1->v2:  Change failure to warning instead of commenting out check.

This patch issues a warning instead of failing when the ipsec configuration
verification check fails so that the network tests don't abort completely.
We need to port the ipsec tests and setup from racoon to openswan in order
to get them to work.

Also removed the "--" from the runcon command line, since it is
no longer supported.

With these changes some of the tests pass but many fail or
get errors.  Also, the tests currently have to be run in
SELinux permissive mode.

Signed-off-by: Linda Knippers <linda.knippers@hp.com>