Updated patch for trustedprograms in base mode
Resending previous patch which now improves exit_fail logging and removes
obsolete TODO comment. The patch is to update failing trustedprograms tests
in Base/CAPP mode. They were failing mostly due to new types of records
and 6.1 improvements.

It respects the old format of tests which are passing on RHEL 5.6,
this could be removed if it is not desired for some reason.