Change expected errno for failed ptrace calls
RHEL6 sets errno to EACCES instead of EPERM when a ptrace attach fails
due to the MLS constraints.  Adjust the tests accordingly.

BTW, I'm not sure why/when the errno changed.  I can see the RHEL6
code paths that retern EACCES but it looks like RHEL5 ought to do
the same, but doesn't.

Also, I noticed the ptrace failure tests can sometimes fail
with a missing audit record if the audit record isn't in the
log quickly enough.  We should probably retry looking for the
record like we do in some other test cases.

Signed-off-by: Linda Knippers <linda.knippers@hp.com>