Don't source testcase.bash in other sourced files
I discovered when running the crypto tests that the top
level directory was left with 2 or 4 tmp files for every test
case that ran.  That's because the test cases are sourcing
testcast.bash, which creates tmp files expecting them to be
cleaned up later, and the helper function files sourced by the
test cases were also sourcing testcase.bash, causing the first
set of tmp files to be forgotten.  Sourcing testcase.bash once
is enough so let's let the test cases do it.

I suppose testcase.bash could be a bit smarter and protect
itself but this works.

Signed-off-by: Linda Knippers <linda.knippers@hp.com>