Remove unnecessary mount check
Our current kickstart script creates a separate file system
for /var/log/audit.  Not sure why but we had a check for that
in our log-related test cases which caused them all to abort.
We don't seem to need this check as the tests work fine even
if /var/log/audit is a separate file system.