Remove unneeded RHEL ipc hack
RHEL5 had non-standard ipc.h file locations so we had
special code in some places.  I removed some of it a
while back but hadn't noticed it in the 32-bit syscall
tests.  With this changes the 32-bit ipc syscalls run,
although the failure cases still fail due to the current
audit bug.

Signed-off-by: Linda Knippers <linda.knippers@hp.com>