rules and policy for labeled networking
Add netlabel rules to automatically label unlabeled
traffic and test policy to go along with that.  I'm not sure
all the policy is necessary but its what I'm running so I'll
add it for now.

With these rules and this policy, I'm able to use ssh
from a non-mls host.  We could certainly tighten these rules
to only label traffic from specific IP addresses or ports
if we need to.  It would just mean each tester would have to
customize his/her configuration.

Also bumped the policy revision to use all the digits.

Signed-off-by: Linda Knippers <ljk@bluefish.(none)>