Makefile to turn on netfilter/ebtables tests
From Jim:

This patch is the first in a series of 9 patches. There are 7 new files
and 2 patched files. The README.netfilter file,    config-server.bash,
and profile.sample file are used by both iptables/ip6tables and ebtables
testing. This set of patches adds the ebtables test functionality.
Hopefully the iptables/ip6tables patches will follow within a week or
so. The tests have been against Intel based platforms running in capp
and mls. One note is that these have been run with the lblnet_tst_server
manually started and not with lblnet_tst-tcp installed as /etc/xinetd.d.
I believe there are still issues with it running in that manner under
RHEL6.1 These test cases have been tested on RHEL6.1 GA  loaded with the
kickstart ks-x86_64.cfg version 0.10-1.

Linda changed the original patch to put the test directory in the
list that only get executed in native mode for the OS.