README.netfilter and config-server.bash update
This patch accommodates the change of the config-server.bash file from
the top-level directory to the utils/netfilter directory in the audit-tests
suite. It also names the profile file created by the script to
profile.<hostname> and looks for the profile.sample file in the
utils/netfilter directory as well.

These patches are to the README file for the netfiltering. The changes
incorporate the fact that the config-server.bash has moved to the
utils/netfilter directory as well as the sample profile and the created
profile.