Script to start lblnet_tst_server
Since I was having a lot of trouble with xinetd working properly both
locally and remote, I started it up manually on the remote end and then
started and stopped it after each test on the TOE that had host= local
in the run.conf (by calling this script).  It seems the main problem I'm
having when trying to run it via xinetd is the commands to kickoff udp
activity never seem to get executed. Running it
standalone works fine (as long as you insure a xinetd isn't running that
has latched onto port 4000).