Fix ebtables ipv6 address setup
This patch corrects an error I found recently where an ipv6 link local
address and an ipv6 global address are assigned to the same ethernet
device (happens often at least on our systems) and the tester wants to
use the global ipv6 address for both the forwarding tests and the
LBLNET_SVR_IPV6. I had (erroneously) assumed in the past that the
LBLNET_SVR_IPV6 would always be the link local address and thus have a
%<interface> scope tacked on to the address used when nc command was run
from the tstsvr_lock routine. This corrects that issue so that if
someone picks the global ipv6 address to be used for everything it
works. This was already taken care of in the netfilter
(iptables/ip6tables) run.conf.