Add tests for new x86_64 accept4() and recvmmsg() network syscalls.
This patch adds tests for the new x86_64 accept4() and recvmmsg()
network syscalls.

With this patch, all accept4 tests will fail due to a bug in auditctl.
I did my testing with an ugly hack to work around this bug. If auditctl
can't be fixed, I'll submit that workaround in a separate follow-up patch.

Signed-off-by: Tony Ernst <tee@sgi.com>