Changes required to get accept4 and recvmmsg tests to pass
This patch has changes to the network/run.conf file to get
the accept4 and recvmmsg tests to pass in both 32 and 64 bit
mode. It is required due to the way RHEL6 works.

I'm submitting it so I don't lose these changes, but if/when
these things are fixed in RHEL6, it can be backed out.

Signed-off-by: Tony Ernst <tee@sgi.com>