Remove strong rnd requirement for non-ssh tests
We have tests that use ssh simply as a tool for testing
something else.  The strong rnd configuration for ssh causes
these tests to take a long time and often fail with timeouts.

This patch updates the trustedprograms/database tests to disable
the strong rng requirement for ssh using functions previously
used by the crypto tests.  As part of this, I'm moving the
tp_ssh_functions.bash file to a more globally accessible
location.

Signed-off-by: Linda Knippers <linda.knippers@hp.com>