Add misc/mount tests
The mount tests in the LTP don't actually seem to test what
we want so we've added a script to run them with a loopback
block device, as we did in our RHEL5 evaluation.  Since these
misc tests use the LTP binaries, we're putting them in the LTP
directory.

Also providing a fix to the mount02 test as it seems to have been broken
in the last year or so.

Merged Santhosh's patches up to the latest git tree.

Signed-off-by: santhosh <santhosh.tn@hp.com>
Signed-off-by: Linda Knippers <linda.knippers@hp.com>