Trigger SA creation for ipsec udp network tests before they run.
Since udp is a connectionless protocol, the SAs are not established
until the first data is written. But for the traffic to go through
ipsec, the SAs need to be established before the socket()/connect().

This patch changes the udp tests to do a preliminary write to the
socket to trigger SA creation before the test runs.

Signed-off-by: Tony Ernst <tee@sgi.com>