Fix account cleanup
These tests create user accounts with 'useradd -Z' but
didn't use the -Z flag when deleting them.  This left
the logins in the semanage information but without
passwd entries, causing semanage to issue warnings.
Now we delete them the same way they are created so
all the information is cleaned up.

Signed-off-by: Linda Knippers <linda.knippers@hp.com>