Randomize the test user and test admin passwords.
The passwords for test user and test admin are hardcoded in
utils/run.bash which makes the passwd fail for any other
than first try as on SLES you are not allowed to set the
password to the same string for second time. Fix that by
appending a random number after the password.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>