Rename `audit' to `audit-test', matching make-dist structure
Previously, "make dist" would basically:
- recurse into "audit", running "dist", creating "audit-test-$rev.tar.gz"
- recurse into "ltp", running "dist", creating "ltp-$rev.tar.gz"
- extract those two to a temp location and create a new archive called
  "audit-$rev.tar.gz", containing "ltp" and "audit-test" directories,
  along with Makefile and rules.mk

This caused inconsistencies between the git repository
and the dist package.

Since the official way to run the suite is to use the dist package,
a simple rename of "audit" to "audit-test" maintains backwards
compatibility while fixing the inconsistency.
The final dist archive name ("audit") also shouldn't cause possible
confusions as of now - it's not just for the "audit" subdirectory.

This change allows runs and development directly on a git working tree.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>