netfilter: use augrok instead of ausearch for tests 45-58
During RHEL6.4 CC retention testing we found that ausearch is run
before required AVC msgs appear in audit log. This is because the
expected AVCs are still in auditd backlog. Using augrok which has
the ability to wait for messages in backlog we cannot reproduce
the issue anymore.

We might do a RFE patch in the future that will replace all ausearch
calls to augrok to mitigate issues like these.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>