kvm-cgroups: fix for libvirt >= 1.1.1 with systemd >= 205
In RHEL7 we will be using the latest versions of libvirt and
systemd. This patch fixes the test scenarios according to the
layout described at the upstream page:
http://libvirt.org/cgroups.html

In RHEL7 the cgroup root folder had changed to /sys/fs/cgroup.
Also by default the ramdisk module (rd, brd) is not loaded
so the tests to pass you need to modprobe brd. This had been filed
under BZ#908050 and this testcase will need to be updated if ramdisk
not loaded is the new default behavior.

Also some small enhancements have been incorporated:
1. remove double destroy of VMs
2. add checks for required kernel modules (kvm, brd)
3. add loading ramdisk kernel module if not present
   (this is disabled in RHEL7 by default)

The brd module needs to be loaded before the code
in functions_cgroup_device.bash.

If no virtual machines were started, the machine subfolder
in cgroups will not exist - do not check for it.

These changes are incompatible with RHEL6.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>