network: swap local<->remote in addr_loop and NS-side configs
The original addr_loop code fills in LBLNET_SVR_IPV[46] with
IP addresses of the TOE(s) just so it can refer to NS as "local"
and TOE as "remote" from the NS's point of view.
This would be common in general networking, but with variables
like LOCAL_IPV4 and LBLNET_SVR_IPV4, the rest of the suite refers
to TOE as "local" and NS as "remote", making this piece of code
very confusing.

This also confused the author of commit 2f8f277, the code
  local_ipv4="$(get_ipv4_addr)"
  remote_ipv4="$LBLNET_SVR_IPV4"
looked innocently enough (in the context explained above), but
since remote_ipv4 contained TOE addresses from client_list.txt,
local_ipv4 had to hold the address of the NS, retrieved here from
"ip addr". Commit 2f8f277 breaks this, because it assumes that
local_ipv4 is always local to TOE.

The original code didn't even work without having LBLNET_PREFIX_IPV6
already exported in the current shell (as enforced by
install_check_server, which is now removed), which is where
addr_filter got local_ipv6 from on the NS.

This commit therefore swaps local/remote in addr_loop and config
files, so that "local" is always TOE-side and "remote" is always
NS-side.

Since the "ip addr" is no longer used (it had its own problems,
described in commit msg of 2f8f277), the address of the NS has
to be retrieved from elsewhere.
Since the user is responsible for assigning IP addresses
to interfaces and since there's no reliable way of getting the
desired address (any interface can have multiple addresses),
the user has to set the addresses manually, hence the README change.
These addresses need to be available during install_server,
so the profile.bash sourcing has been added.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>