add new sanity-checking script, envcheck
The idea is to catch various configuration or setup errors
and give useful hints, so that the user doesn't have to debug
the system/suite to find out what could possibly go wrong.

This script is not supposed to replace or duplicate functionality
provided by the suite itself, it should be only used for basic sanity
verification of the environment. It doesn't claim to reveal all
configuration errors, but it can still be useful.

Because of the reasons mentioned above, running this script
is purely optional.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>