implement run/rollup log result merging
This patch incorporates changes that make possible
merging of subsequent test runs.

The logs are now stored also separately for each test
in the logs subdirectory of the test bucket. The run and
rollup logs are generated from these logs. The output
of the run.bash should remain the same.

This patch also fixes the --list option that now does
not delete the contents of run and rollup log.

This patch adds generate (-g|--generate) option that generates
the total run and rollup logs. This might be useful if these
were deleted or not created after aborting testing with SIGINT.

The distclean make target removes the new logs directory.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>