audit-remote,fail-safe: /var/log/audit on tmpfs for private namespaces
In RHEL7 the concept of PAM namespaces was introduced with the / mount point
being mounted as private by default CC evaluated configuration. This way any
mount point mounted by an user X cannot be seen by the other users (the only
exception is when X is appropriately excluded in PAM namespace configuration).

Therefore we have to make sure that /var/log/audit is mounted by eal so that
eal user is excluded from PAM namespace during the test and the exclusion is
removed afterwards.

Signed-off-by: Ondrej Moris <omoris@redhat.com>