utils/network-server: misc fixes, new "detach" command
This patch adds a new "detach" command, which can be used to close
the connection with the remote (controlling) client. This is useful
for executing commands "in the background" on the server, ie.

  "echo:server_works;detach;sleep:5;recv:ipv4,tcp,1234,0;"

which would return "server_works" to the client and then close
the connection, doing recv after 5 second sleep.

The implementation of this command required several additional code
changes / fixes, which should be pretty minor.
Those include:

  - `sock' -> `*sock' typo fix
  - memory leak fix due to fdopen usage
  - *_hlp_* function robustness (`-1' fd value)
  - possible segfault when calling shutdown on uninitialized ctl_sock

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>