utils/network-server: server-side cleanup on port 4009
This mechanism provides a generic way of cleaning up any possible
lblnet_tst_server instances that might be running, and,
due to "instances = 1" xinetd option, blocking the execution of new
instances.

The usage is as simple as

  nc <addr> 4009

or

  ncat <addr> 4009 < /dev/null

to make ncat quit immediately after opening the connection
(which is enough to spawn the cleanup tool by xinetd on the server).

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>