audit-remote: Added a waiting time when a message is sent via audit-remote.
Recently audit has changed in a way that it does not empty remote queue
when it is stopped. Our tests using (in a quite non-standard way) audisp-remote
to sent message to the remote end were failing due to this change, adding an
appropriate wait time make sure that testing messages will be sent before
audisp-remote terminates.

Signed-off-by: Ondrej Moris <omoris@redhat.com>