network: properly test timeout on connect() expres=fail
Original code *relies* on the server giving up sooner than the client:

  1. "recv" operation on port 4300 is requested from server
  2. client tries to connect() with an invalid context, packets get
     dropped on the server side (no replies sent)
  3. server times out after 10 seconds (original -t 10 in xinetd conf)
  4. when client sends another TCP SYN (conn request), it hits a closed
     port and the server replies with TCP RST, which connect()
     interprets as ECONNREFUSED, passing the test

This is an incorrect approach, because it doesn't verify the packet
droping behavior, only that the connection didn't succeed. If, for some
unexpected reason, the remote socket never started listening, the test
would pass anyway.

Furthermore, it presumes that the client continues connecting even
after the server times out, which is *not* guaranteed and - in fact -
is the reason that inspired this patch.

The new approach is based on the fact that the testing machine now
controls both start and exit of the test server (after recent changes),
making it possible to keep the remote "recv" operation active until
the expres=fail tests properly time out, verifying the correct behavior.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>