tp/lpq: disable pam_loginuid
In recent RHEL7 kernel changes the loginuid call fails with
pam_loginuid where loginuid already set in session. This means
that all test that expect login will fail in pam with message:
Cannot make/remove an entry for the specified session

This fix temporarily disables pam_loginuid in /etc/pam.d/login
as a workaround. This has no effect on testing purpose of
this test.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>