tp/lpq: fix for RHEL7 cups
By default cups shows owner and job name only for jobs
owned by the user. If the user is in the system groups
sys or root he will be able to see this information for
all available jobs.

This patch makes the test user, for the time of the test
run, a member of sys group so he can correctly see the
job names needed to verify correct SELinux enforcement
in cups.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>