libpam/loginuid: test pam_loginuid functionality
This test checks if pam configurations for all entry point applications
include pam_loginuid in session management group.

To test the correct functionality of pam_loginuid ssh and login is used.
The login via login command should fail as login uid is already set in
the session. This is a new security feature introduced in this kernel
commit [1] and is available in RHEL7. When connecting via ssh the login
should succeed and correct USER_* events should be logged in audit.log.

[1] https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=633b4545

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>