utils/functions.bash: fix start_service for cups
This patch extends start_service to support for cups startup.
It uses busy waiting on the service start detected by creation of
cups socket - /var/run/cups/cups.sock.

The motivation of this change was to fix issues with
trustedprograms/cupsd test which could fail while searching
for a specific audit event, because cups did not yet
properly start and did not emit the event - i.e. race condition.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>