network: remove incorrect/useless getaddress functionality
This feature was useful back when the suite was hard-locked to using
only the first NIC as lblnet. This limitation has been lifted and
the interface is now selectable by the user (using LOCAL_* variables),
which means that "getaddress" can no longer automatically guess what
interface or addresses are going to be used (by design).

The current behavior is to just echo what the user provided in
LOCAL_IPV* variables. Furthermore, the Makefile target depends on
install_check, which makes the user set LBLNET_SVR_IPV* as well.

In conclusion, the getaddress is now simply useless as it prints out
already known information, entered by the user. Its usage, as
recommended by README.run, is also invalid as it fails on
LBLNET_SVR_IPV4 being unset at that time.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>