misc/procperm: don't fail when removing nonexistent group
The group may have been removed by userdel due to 'USERGROUPS_ENAB yes'
being set in login.defs.

If - in delete_user - the last groupdel fails, its non-0
return value is carried as a function return value back
to the parent shell and is evaluated as ERROR, so return 0
explicitly.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>