utils/bin/do_fork: remove getrlimit stderr output
The output was conflicting with "testres exitval pid" fprintf
and doesn't seem to be used anywhere in the suite.
Furthermore, it doesn't alter the behavior of the fork syscall,
since it's getrlimit, not setrlimit.

Whitespace separation from the rest of the code suggests that this
was perhaps a debugging (temporary) code.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>