rules.mk: empty the 'verify' target
The mount options verified by this target are highly fs-specific
and relevant only to ext* filesystems. This target therefore fails
on ie. RHEL7, which uses xfs.

Even on ext* filesystems, these options don't have to be specified
as mount options, in fact the *defaults* for mke2fs are to enable them
in the metadata (see tune2fs -l /dev/sdX | grep 'Default mount').

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>