audit-tools: new test for audit rules
It should be possible to select the set of events to be audited from the
set of all auditable events based on various attributes such as, for
instance, a type of audit event, subject or user identity, outcome of the
audit event and so on. Newly added "rules" test in audit-tools bucket
performs testing of this functionality.

Signed-off-by: Ondrej Moris <omoris@redhat.com>