utils/bin: remove do_send
The current version was built only on ppc64 where the send(2) syscall
actually exists. No tests use it, the MSG_OOB/URG functionality was
copied to do_sendto (in a more backwards-compatible way).

If - for some reason - you need it, please create a copy of do_sendto
and replace the actual sendto() with syscall(__NR_send, ..).

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>