README.run: remove obsolete workarounds
network:
This section describes a generic way to run tests manually, which is
already described in sections above. The same steps (except for
./run.bash instead of `make run') are used for regular execution since
RHEL7 versions of the suite.

fail-safe:
These tests are no longer expected to fail on RHEL7.

trustedprograms:
The aide_update test runs in both capp and lspp and should work now.

password-on-NS:
No longer relevant as RHEL7 uses systemd and there's no run_init
or PASSWD required on NS.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>