utils: implement test runtime detection / logging
This introduces a "test time" counter for each bucket - duration of
the tests only, without run.bash setup. May be useful for performance
regressions reasons as well as for simple user information.

Runs with 'make rerun' are accumulated into this per-bucket runtime.

Note that the current implementation has only second resolution, meaning
if all tests together in a bucket manage to finish in <1sec, the total
time will be 0s. This doesn't normally happen, but may occur on rerun,
in which case 0sec will be added to the cumulative counter.

Regarding printout and logs,

 - console needs values relevant to the current run/rerun
 - log needs total values (relevant to all runs of the bucket)

meaning that the displayed pass/fail/error/total/runtime values should
match the current/last (re)run, but they need to have absolute values
in the rollup log.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>