audit-remote: active waiting for replies from NS added
On non-x86_64 system, responses from NS might be delayed, moreover
responses can be delayed by network on any system. Previously 2
seconds sleeps were used to wait for replies but sometimes it is
not enough. Now we check for replies every 2 seconds, once we got
it, waiting is terminated with success. After 5 rounds (~10 seconds)
waiting is terminated and failed.

Signed-off-by: Ondrej Moris <omoris@redhat.com>