misc: make cgroups_limits binary wait after sending freezer2
The previous code had a race condition - the cgroups_limits binary could
send "freezer2" and exit before the test (parent) could read it. This
can happen if a process is allowed to write(2) to stdout and the
write(2) doesn't block because of kernel buffers.

The solution is to add another synchronization wait after sending the
second ("freezer2") string.

Also signal the cgroups_limits process itself, not timeout(1). The same
code can be used for FROZEN case because the cgroup_exec binary replaces
itself with cgroup_limits and in both cases we obtain the pid of
cgroup_limits after reading "freezer1" to make sure timeout(1) or
cgroup_exec had enough time to spawn cgroup_limits.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>