tp/labeled_net: add sshd labeled networking tests
This patch adds additional two scenarios to the test:

1. Socket activated, labeled networking aware sshd spawned by systemd
on port 222. The security level of the logged in user must correspond
to the security level of the incomming connection. It cannot be overriden.
Note that this test scenario requires "allow init_t lo_netif_t:netif ingress;"
SELinux rule to work - included # via lspp_test policy. This policy is
not included in the default policy as it is not an usual scenario and is
useful only for testing.

2. Login at different security level via "Traditional" SSH instance running
on port 22. This is not related to labeled networking in any way, but is really
close in testing scenario thus is included here.

It also adds a new ssh_cmd_mls_user helper function required by both
scenarious.

This is a reproducer for these two bugs found in RHEL7.1:
https://bugzilla.redhat.com/show_bug.cgi?id=1201381
https://bugzilla.redhat.com/show_bug.cgi?id=1202843

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>