utils/tp_ssh_functions: addressing robustness of ssh_mls_cmd
On machines with more than one processor, race conditions in expect
script might happen in specific circumstances (which depend on, for
instance, kernel scheduler tuning). For instance, expect is sending
as soon as it match a desired output of an application but the
application is not yet listening. Expect scripts running on such
machines need more robustness - in particular, "send" commands
delayed by "after" command.

Signed-off-by: Ondrej Moris <omoris@redhat.com>