audit-tools/rules: mitigate race conditions and speedup
On overloaded s390x guest this test was quite unstable
due to possible race conditions. This patch tries to mitigate them.

Use small sleep after adding an audit rule before executing
the reproducer. This mitigate possibility that the rule is applied
before the reproducer is run.

Use an active wait when looking for audit messages up to 5s. In some
cases on our s390x waiting for backlog was not sufficient.

This test now uses audit.log cleanup to have a clear audit.log
before each test. Using -ts implied the need to add 1s sleeps
to achive this.

The test now runs a lot faster and looks finally stable.
On our s390x machine it passed 100x in a row.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>