crypto: a potential race condition issue fixed in TLS test
Previously it might happen that TLS server used during the testing was
killed prematurely. After connecting NSS tstclient to TLS server, there
is a waiting for a right moment to return from tls_client_connect function,
then TLS server is killed and the client terminates afterwards. However
NSS tstclient is sometimes expected to fail and sometimes expected to pass.
Thus the waiting is terminated with messages

 * "SSL_ERROR_NO_CYPHER_OVERLAP" (expecting FAIL), or
 * "Read from server 0 bytes" (expecting PASS).

There was also a specific waiting case when something unexpected happened

 * "PR_Poll returned 0x00"

since the NSS tstclient is running in the background, the last case could
be potentially matched before matching either FAIL or PASS messages if the
tstclient was frozen for a moment.

Signed-off-by: Ondrej Moris <omoris@redhat.com>