tp/tp_selinux_functions: enhance load/update policy funcs
Add support for >= policycoreutils-2.5. The logic for naming
the policy modules has changed. The name of the module is now
taken from the file name and the policy_module directive is
ignored. Add support for loading and updating a renamed policy
module instead of the original file name.

This is the least intrusive solution to make the semodule test
to work with >= RHEL7.3 and Fedora.

This change is backward compatible.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>