audit-tools: minor generalization in permissions test
File audit.rules has by default "r" right for a group. However,
whenever this file is deleted and autogenerated from scratch, it
has no longer "r" for a group (because mktemp is used during its
autogeneration). Since correct group is always "root" this change
does not matter. Test checking correct permissions on this file
now passes in both cases (with or without "r" for a group).

Signed-off-by: Ondrej Moris <omoris@redhat.com>