trustedprograms: time issue fixed in account_lifecycle test
Test account_lifecycle triggers various audit event for user and group account
lifecycle and verifies that event are correctly logged. On slower architectures
like ppc64 it might take a second to drop an event to the log file. Previous
version of the test was checking log files right after audit backlog was empty
which was not always enough on aforementioned slower architectures. Updated
version of the test does not rely on backlog status but waits for events to
appear in log file via wait_for_cmd function.

Signed-off-by: Ondrej Moris <omoris@redhat.com>