utils/ns2/ipa: do kinit always, before host-show
The issue was that host-show here is meant to detect if a host
is missing in the IPA DB, as host-add unconditionally follows.
The problem is that if the current user doesn't have a krb5 ticket,
host-show also fails even though the host is already in the DB.
As the kinit was done just before host-add, this made the host-add
request valid, but adding already existing host, which failed.

The solution is to simply do kinit always, globally, like other
foreign ipa/sssd testing scripts do. This doesn't guarantee that
host-show will fail only on nonexisting host, but it makes it much
more likely.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>